This commit was manufactured by cvs2git to create tag 'EventFilter- StorageManager-V05-02-03'.
Sprout from master 2009-10-16 13:34:32 UTC Remi Mommsen <sha1-09ace11b042676bf0011c424f6ae14c5eceb9645@cern.ch> 'Prepare for tag V05-02-03.'
Cherrypick from master 2009-09-29 15:29:59 UTC dshpakov <sha1-82b916591e9d6901e66a09082a9b194c199c9987@cern.ch> 'Moving ErrorEventData':
    src/ChainData.cc
    src/ChainData.h
Delete:
    src/EndLumiSectMsgData.cc

+++ src/ChainData.cc
@@ -1,4 +1,4 @@
-// $Id: ChainData.cc,v 1.1.2.2 2009/10/13 14:13:54 mommsen Exp $
+// $Id: $
 
 #include "toolbox/mem/Reference.h"
 
@@ -44,13 +44,11 @@ detail::ChainData::ChainData(toolbox::mem::Reference* pRef) :
   // variable default value for one of the fragKey fields.
   if (pRef)
     {
-      _fragKey.secondaryId_ = static_cast<uint32>(
-        (uintptr_t)pRef->getDataLocation()
-      );
+      _fragKey.secondaryId_ = (uint32) pRef->getDataLocation();
     }
   else
     {
-        _fragKey.secondaryId_ = static_cast<uint32>( time(0) );
+      _fragKey.secondaryId_ = (uint32) time(0);
     }
 
   if (pRef)
